refactor: устранить все 3 циркулярные зависимости
Проблема:
- 3 циркулярные зависимости в проекте:
  1. delete-attachment.definition.ts ↔ delete-attachment.tool.ts
  2. search-tools.definition.ts ↔ search-tools.tool.ts
  3. base-definition.ts ↔ tool-metadata.ts

Решение:
- Применен паттерн "Extract Metadata" для разрыва циркулярностей
- Создан delete-attachment.metadata.ts с метаданными
- Создан search-tools.metadata.ts с метаданными
- Создан base.types.ts для общих типов

Изменения:
1. yandex-tracker/delete-attachment:
   - Definition и Tool теперь импортируют только metadata
   - Убраны взаимные импорты между definition и tool

2. framework/search/search-tools:
   - Аналогичный рефакторинг с metadata файлом
   - Убран дубликат описания tool

3. framework/core/base:
   - Создан base.types.ts для ToolDefinition
   - base-definition и tool-metadata импортируют из base.types
   - Разорвана type-level циркулярность

4. Исправлен issue-url.tool.ts:
   - Убран async без await (заменен на Promise.resolve)
   - Исправлена ошибка компиляции TypeScript

Результат:
✅ 0 циркулярных зависимостей (было 3)
✅ lint, typecheck, build успешно
✅ cpd: 3.22% (ниже порога 5%)
✅ depcruise: 0 нарушений

План: .agentic-planning/plan_fix_validation_issues/
2.1_refactor_circular_deps_parallel.md
